refactor: move withAPollo to api-client
