+++ src/pages/index.tsx
@@ -4,7 +4,7 @@ import Sidebar from '../components/Sidebar';
 import VisuallyHidden from '../components/VisuallyHidden';
 import { useMapState } from '../components/MapState';
 import config from '../config';
-import { withApollo } from '../components/withApollo';
+import { withApollo } from '../api-client/withApollo';
 import LooMap from '../components/LooMap/LooMapLoader';
 
 const SIDEBAR_BOTTOM_MARGIN = 32;
